third_party/blink/web_tests/webaudio/internals/cycle-connection-gc-expected.txt

Cycles of AudioNode connections should be collected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

A cycle was created:
PASS internals.audioHandlerCount() > initialCount is true
GC happened:
PASS internals.audioHandlerCount() is initialCount
PASS successfullyParsed is true

TEST COMPLETE
